Added a high level pipeline control script
Instead of a high level controlling script for a specific
observation and map, we should have a robust and generic high
level control script that takes observation specific variables
as input parameters.  This script is the first step in that
direction, and at the moment is simply a copy of a specific
observation mapping script.