Added script to find all mapping blocks in an SDFITS file.
This script will list ranges (begin,end) for blocks of mapping
scans.  This included both RALongMap and DecLatMap scans.  The
ranges are by default echoed to the screen and also written to
a file in the /tmp/ directory.

It is not connected to the pipeline, but could be used as a wrapper
around it to itentify start and stop mapping scans, then invoke
the pipeline for each map block.  The other input parameters are
largely dependent upon the observation calibration strategy.