Use sdfits file name to set idlToSdfits outfile name argument.
The default name was just the source name, the new default is to use the
output sdfits file name if it follows the pattern used by the pipeline
(otherwise the previous behavior of letting idlToSdfits pick the output
name is used).

Also fixed a bug when vSourceEnd was not specified.