Modified scripts for Green Bank installation.
There are some ugly hard-coded paths now, but this allows the code
to be run from anywhere in Green Bank (in theory).  All that should
be required is to add /home/sandboxes/kfpa_pipeline to your
PATH and PYTHONPATH environment variables.  Also, the hard-coded
path are a temporary fix while we still have IDL code in the pipeline.

After the transition to python, all code can be imported as modules.