Added logging of many simultaneous pipes (part 1 of 2).
All of the simultaneous pipes are triggered by the setting
of --allmaps when run on an input dataset with multiple
mapping blocks.  General common info is written to the screen.
Othere detailed information for each map is written to a log
file named by the scan number range covered by the map.