Add feed and polarization selection, sequentialize imaging.
The user can now give an inclusive list of feeds and/or
polarizations to filter what is processed in a single map.

These options are not currently to be used with --allmaps.
This implementation only supports an explicitly defined map
using the -b -e and --refscan parameters.

Example usage of the new parameters:
  -f 2,3,4
  -p 'RR'

Also, imaging with AIPS is sequentialized per map because
only a single AIPS id is used and multiple AIPS sessions
for each map with a single user id may cause problems.