Disabled scale with map scans, by default.  Bug fixes.
--no-map-scans-for-scale is replaced with
--map-scans-for-scale with the opposite behavior.  This
effectively just changed the default to not use map scans
for scaling and should allow maps without noise diode firing
to be calibrated.  It should also allow the pipeline to run
a little bit faster.

Now a range of one scan can be defined (ie. '5:5') without
failing.  Also, the pipeline will no longer fail when a
reference scan that is not the same was what is automally found
using the index file.