Corrected the computation of Ta* without aperture efficiency.
And generally cleaned up the aperture and main beam efficiency
code.  A couple of hard-coded values should be exposed to the
command line, including the Ruze equation parameter (tab. 2,
Langston)and the GBT-specific gain constant which accounts for
the Boltzmann constant and the collecting area of the telescope
(eqn. 4, Braatz).