Got parallel sampler calibration going again.
This had been dropped in favor of parallel map calibration.  It
was supposed to be temporary, but stayed on longer than expected.
Imaging seems to be the bottle neck now.  Parallel imaging of
each sampler should be a near-term target.  Also, calibration on
the next map could start while the previous is being imaged, but
this complicates the flow of the pipeline.