Added several new options, noise diode check and low freq opacity.
The user can now set the 0 frequency channel main beam efficiency
so that etaMB is computed seperately from aperture efficiency, etaA.

The user can also for a FS map to be calibrated as PS.

The default zenith opacity for frequencies < 2 GHz is now to
to .008.

There is a new check to make sure the noise diode is firing for
the refscans.