Updated to use system Obit installation, ParselTongue2, old AIPS.
ParselTongue2 can be used with the system-side Obit install,
although it uses Python2.4.  I removed the pyfits lines from the
imaging scripts because they were not needed in the contrib
python scripts.  The AIPS sqash task revealed a but after the new
year change to a new version.  I changed the doImage script to use
the OLD AIPS version for stability and to aviod this problem.