Updated to allow sdfits to run on non-acs data, added doctest for tatm.
If necessary, the pipeline can be used to invoke the sdfits filler when
SDFITS input is not available.  Previously, when the filler wass run, it explicitly
set backend=acs. This was updated to allow vegas (and any other) SDFITS files to be
filled as well.

We could make it more sophisticated later, but what I just coded up was to not
use -backends, then after sdfits is done look to see if there is more than one
output SDFITS file.  If so, there is an error and the user is asked to check
the input directory and provide only one input to the pipeline.