Removed explicit backend id from infile check. Added doctests.
Previously the input fits file had to end in acs.fits or vegas.fits,
but now it just eneds to end in fits.

A couple comments were updated.  An error message was updated.
An imported and unused module (multiprocessing) was removed from
gbt_pipeline.py.  Doctests were added to a couple functions in
pipeutils.