Initial commit of non-mapping pipeline code.
This prototype code operates independently from gbtpipeline.  It is
designed specifically for GBT extragalactic H1 OnOff or OffOn
observations as a starting point for a more generalized non-mapping
pipeline.  The final product is a SDFITS formatted spectrum for each
target in an observation.  So, if an input file has many total power
integrations for 12 targets, then the output will be a SDFITS file
with 12 rows, each representing a calibrated spectrum for a single
target.

The quick-look pipeline features RFI excision using two rudimentary
methods:

1) We examine each integration for broadband RFI (excessive
spectral baseline ripples) and flag integrations based on the total
power low-order Fourier components.
2) We flag narrow-band RFI using an iterative median filter approach.