Fixed two bugs: bad velocity axis in FITS output, fail on bad procedure.
The velocity axis bug was the result of rebinned data no being
consistent with the necessary header keys.  The fix was to disable
rebinning in favor of higher resolution output.

The bad procedure bug was that the code assumed either OnOff or OffOn
procedures in the input for every target, even if not for every scan.
The fix is to no longer make this assumption.