Bug fix to not fail on imaging if AIPS setup was already run.
The imaging step would fail when the AIPS login script
(e.g. . /home/AIPS/LOGIN.SH) was already run in the user's
environment.  It turns out that AIPS would not set the VERSION
environment variable if it was already set, which was being
done by the pipeline.  I simply changed the variable the pipeline
script used to GBTPIPEVERSION.

I also added code to Imaging.py to allow local imaging steps to be
run, mainly for development.