Fixed bug for older data where user set velocity to zero.
A lot of older data was observed for redshifted galaxies and the
observer set the rest frequency to the desired sky freq, and velocity
to zero.  The pipeline crashed on this kind of data because
sourceType does not get set.

Now we set sourceType to "TP_Source" by default.  We still set to "HI"
if the rest frequency is close to 1420 MHz.

If we detect sourceType = "TP_source" then process as usual for HI,
but skip the flagging near zero velocity that would otherwise blank
out the GC HI emission.