Overhauled the dbcon.py script contributed by Glen.
This is hopefully much easier to read now. It should also run more
cleanly and efficiently. One notable difference in usage is that the
script will not clear the AIPS catalog anymore unless the user gives
permission: either by an interactive prompt or by setting a command
line flag (--empty-catalog).  Another difference is that this script
no longer writes out a DBCON fits file.  That was not a necessary
feature to create an image.

NB: a couple other scripts like mapDefault may need minor modification
related to the state of the AIPS catalog at the end of this script.
However, the may be fine as is -- even though they are likely to be
updated soon anyway.