Updated to latest fitsio library to fix COMMENT bug.
Bob noticed an issue where some header COMMENTS came out a little
different in the output file than in the input.  Apparently, this
caused CASA to crash, which is probably also a bug in CASA.  Now,
the COMMENTS are duplicated in this latest update, so it's still
not perfect but it may be enough to make CASA happy and be OK
for now.