Log errors when opacity not in DB.
I found a problem where errors were not logged when the
zenith opacity could not be found because the observation date
was too early.

I also cleaned up an error when tyring to remove the summary/
directory when it wasn't empty.

Finally, I updated a couple method names for the newer version
of fitsio.