Bugfix for shift in freq.-switch calibration.
Sometimes the pipeline would use the center frequency
for the reference data instead of the signal, depending
on the order it was written to the raw data file.  This fixes
that by only grabbing the center frequency from the input
FS signal spectrum for each integration.

I also improved the input file summary screen output.