Added info to stats.log of pipeline runs.
In addition to who, when and what was ran, the stats.log file now
also contains the where (i.e. machine name and directory).

I also removed the comma between the pipeline version and command
parameters to make it easier to copy-paste a command from the log
file and run it again.