Added old AIPS scripts with redirect, fixed small imaging bug.
The old aips scripts contrib/dbcon.py and contrib/mapDefault.py
now simply redirect the user to load.py and image.py, respectively.
This is meant to ease the transition from the old scripts to the
new.

I also found a but where AIPS fails on a long input file name. I
updated the code to create a short temporary file name only when
loading the data.