Add early warning with imaging not on arcturus.
This prevents the mysterious no output scenario when running the
pipeline with imaging on a machine other than arcturus with a
different AIPS setup.  Before you would have to look into the log
file and hopefully see a useful AIPS error.  The way the pipeline
is setup, I would otherwise need to expose idlToSdfits "errors",
which are sometimes harmless at the same time as the useful AIPS
errors.  This is one of the most common AIPS error situations.