Updated usage and warnings for old scripts. Fixed load.py bug.
The old scripts were showing usage for the new scripts.  I fixed
this to show old usage and update the warnings to make them more
informative.

I found a problem with AIPS when the filenames were too long.  I
fixed this by making a temporary short name as a symlink.

Load.py had a problem where the order of the DBCON-ed data changed
the output image.  That is now fixed to match the release.