Added scripts to update images to Galactic coordinates.
AIPS only produces images where the header is written with
equatorial coordinates.  If the original data was in galactic
coordinates, the header needs to be updated to be viewed properly
in tools like the casaviewer.  These scripts modify the CTYPE1
and CTYPE2 keywords to have the correct galactic values.  The image
itself is already correct and unaffected by these updates.