Added sdextract script as workaround for >32k chan data.
The idlSdfits format converter code can not handle data with
more than 32k channels.  This sdextract script is a short term
workaround that allows the user to extract sub-spectra and/or
smooth the calibrated spectra and produce smaller input spectra
for idlToSdfits.

A medium-term solution will be a python version of this code that
is callable from within the pipeline.  A longer-term solution is
a replacement for idlToSdfits that can handle an arbitrarily high
number of channels.