UG update: moved scaling factor from Ta* to Tcal equation.
Last week Jim, Bob and I decided the proper place for the per beam
scaling factors is as an adjustment to the Tcal values, which is used
to determine system temperature, Tsys.  The code and documentation is
being updated to reflect this change in calibration.  This commit is
the change to the documentation.

I also added page numbers to the user's guide.