Updated AIPS disk setup to default to lustre.
We have recurring issues with the local arcturus disks
filling up.  This changes the setup to write to lustre
first.  A potential problem is the efficiency of writing
many files to lustre.  We will need to monitor the run time
for a while to see if this is the right solution.